게시물 생성 API 구현

+++ src/main/java/videoservice/domain/adVideo/service/AdPickService.java
@@ -4,5 +4,5 @@ import java.util.List;
 
 public interface AdPickService {
 
-    List<String> pickAdList(int size);
+    List<String> pickAdList(long size);
 }
